Added all Current Skill Descriptions
This folder of txt files is the master list for all skills now. If you want to add a skill, first make a txt file for it. Then add it to the code. If you want to change a skill, change it in the txt file then change all instances of it in the code. Always do both, starting with the  txt. When in doubt, do what the txt file says. No skills with different effects will share a name.